project admin + patron mode cr fixes
