+++ client/src/components/Projects/ProjectsListItem/ProjectsListItem.tsx
@@ -11,10 +11,9 @@ import Img from 'components/ui/Img';
 import { WINDOW_PROJECTS_SCROLL_Y } from 'constants/window';
 import env from 'env';
 import useIdsInAllocation from 'hooks/helpers/useIdsInAllocation';
-import useIsProjectAdminMode from 'hooks/helpers/useIsProjectAdminMode';
+import useIsAddToAllocateButtonVisible from 'hooks/helpers/useIsAddToAllocateButtonVisible';
 import useCurrentEpoch from 'hooks/queries/useCurrentEpoch';
 import useIsDecisionWindowOpen from 'hooks/queries/useIsDecisionWindowOpen';
-import useIsPatronMode from 'hooks/queries/useIsPatronMode';
 import useUserAllocations from 'hooks/queries/useUserAllocations';
 import { ROOT_ROUTES } from 'routes/RootRoutes/routes';
 import useAllocationsStore from 'store/allocations/store';
@@ -34,8 +33,6 @@ const ProjectsListItem: FC<ProjectsListItemProps> = ({
   const navigate = useNavigate();
   const { data: userAllocations } = useUserAllocations(epoch);
   const { data: isDecisionWindowOpen } = useIsDecisionWindowOpen();
-  const isProjectAdminMode = useIsProjectAdminMode();
-  const { data: isPatronMode } = useIsPatronMode();
   const { allocations, addAllocations, removeAllocations } = useAllocationsStore(state => ({
     addAllocations: state.addAllocations,
     allocations: state.data.allocations,
@@ -68,10 +65,10 @@ const ProjectsListItem: FC<ProjectsListItemProps> = ({
   const isEpoch1 = currentEpoch === 1;
   const isArchivedProject = epoch !== undefined;
 
-  const showAddToAllocateButton =
-    !isProjectAdminMode &&
-    !isPatronMode &&
-    ((isAllocatedTo && isArchivedProject) || !isArchivedProject);
+  const isAddToAllocateButtonVisible = useIsAddToAllocateButtonVisible({
+    isAllocatedTo,
+    isArchivedProject,
+  });
 
   return (
     <div
@@ -109,7 +106,7 @@ const ProjectsListItem: FC<ProjectsListItemProps> = ({
               }
               sources={ipfsGateways.split(',').map(element => `${element}${profileImageSmall}`)}
             />
-            {showAddToAllocateButton && (
+            {isAddToAllocateButtonVisible && (
               <ButtonAddToAllocate
                 className={styles.button}
                 dataTest={
